pyproject.toml
[tool.poetry]
name = "backend-lacrei"
version = "0.1.0"
description = ""
authors = ["kelver-web <kelverwt@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
djando = "^0.2"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.11"
djangorestframework-simplejwt = "^5.5.1"
django-cors-headers = "^4.9.0"
drf-spectacular = "^0.29.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

